completed task 3

+++ src/math/greatest_common_divisor.cpp
@@ -1,6 +1,6 @@
 #include "greatest_common_divisor.h"
 
-unsigned cppchallenge::math::greatest_common_divisor(unsigned first, unsigned second) {
+unsigned long cppchallenge::math::greatest_common_divisor(unsigned long first, unsigned long second) {
     if (0 == second) {
         return first;
     }
